+++ cli/web_worker.rs
@@ -37,12 +37,10 @@ impl WebWorker {
       ops::web_worker::init(isolate, &state, &worker.internal_channels.sender);
       ops::worker_host::init(isolate, &state);
       ops::io::init(isolate, &state);
+      ops::resources::init(isolate, &state);
       ops::errors::init(isolate, &state);
       ops::timers::init(isolate, &state);
       ops::fetch::init(isolate, &state);
-      // FIXME(bartlomieju): this is added only to provide "close"
-      // op - it should be moved to `ops::io`
-      ops::files::init(isolate, &state);
     }
 
     Self {
